refactor(codec): split alphanumeric codec helper
